Created file tree
With these steps, I created the intital file tree to begin working from. I added CSS styling to test it, used a Google Font to make sure that I correctly integrated it. I used a test Javascript file to make sure it is working well. I'm happy with this start!