Video and P layout
I got the video added! Then I was able to put it in the correct position that it will be in. Put a paragraph underneath it. Then added on media queries to make the site responsive! Now it works on all screen sizes, doesn't break. Content just stacks and shifts as needed.