NavBar work and map download
I added an Ohio map to show the service area for the window cleaning business.
On the navbar, I didn't like the home page link setup because it had the index file in the media folder. A good file setup is important to me, so I moved the index out of the media folder and then had to update my navbar links