Learning experience!
This evening I decided to spend it doing something I truly enjoy. CODING! Looking things up on MDN, and just making a fun, functional site that is what I envisioned. Done with the index! Had fun, learned several new things. Such as setting a video as a background, making a menu with images and text on it to select.